+++ ghosts.cpp
@@ -34,25 +34,25 @@ std::vector<int8_t> cycleTimes = {
   -1
 };
 
-int8_t blinky_cycle_index = 0;
-int8_t pinky_cycle_index = 0;
-int8_t inky_cycle_index = 0;
-int8_t clyde_cycle_index = 0;
+uint8_t blinky_cycle_index = 0;
+uint8_t pinky_cycle_index = 0;
+uint8_t inky_cycle_index = 0;
+uint8_t clyde_cycle_index = 0;
 
-void manage_ghost_move_state(Timer &t, Ghost *g, int8_t *cycle_index) {
+void manage_ghost_move_state(Timer &t, Ghost *g, uint8_t *cycle_index) {
   ghostState nextState = cycleStates[*cycle_index];
-  printf("%s::manage_ghost_move_state cycle %d setting state %d.\n", g->name.c_str(),  *cycle_index, nextState);
+  //printf("%s::manage_ghost_move_state cycle %d setting state %d.\n", g->name.c_str(),  *cycle_index, nextState);
   g->set_move_state(nextState);
   if (*cycle_index < cycleTimes.size()) {
     int8_t next_cycle_time = cycleTimes[*cycle_index];
     if (next_cycle_time != -1) {
       // Setup next ghost cycle.
-      printf("%s::manage_ghost_move_state timer pre  dur - %d started - %d loops - %d state - %d\n", g->name.c_str(), g->move_cycle_timer.duration, g->move_cycle_timer.started, g->move_cycle_timer.loops, g->move_cycle_timer.state);
+      //printf("%s::manage_ghost_move_state timer pre  dur - %d started - %d loops - %d state - %d\n", g->name.c_str(), g->move_cycle_timer.duration, g->move_cycle_timer.started, g->move_cycle_timer.loops, g->move_cycle_timer.state);
       g->move_cycle_timer.duration = cycleTimes[*cycle_index] * 1000;
-      printf("%s::manage_ghost_move_state timer post dur - %d started - %d loops - %d state - %d\n", g->name.c_str(), g->move_cycle_timer.duration, g->move_cycle_timer.started, g->move_cycle_timer.loops, g->move_cycle_timer.state);
+      //printf("%s::manage_ghost_move_state timer post dur - %d started - %d loops - %d state - %d\n", g->name.c_str(), g->move_cycle_timer.duration, g->move_cycle_timer.started, g->move_cycle_timer.loops, g->move_cycle_timer.state);
       (*cycle_index)++;
     } else {
-      printf("%s::manage_ghost_move_state stopped.\n", g->name.c_str());
+      //printf("%s::manage_ghost_move_state stopped.\n", g->name.c_str());
       g->move_cycle_timer.stop();
     }
   }
@@ -83,7 +83,7 @@ void Ghosts::init(uint8_t current_level) {
 
 void Ghosts::move_start() {
   for (auto ghost : ghosts) {
-    printf("Ghosts::move_start Starting move cycle timer for %s\n", ghost->name.c_str());
+    //printf("Ghosts::move_start Starting move cycle timer for %s\n", ghost->name.c_str());
     ghost->move_cycle_timer.start();
   }
 }
@@ -98,7 +98,7 @@ void Ghosts::move_resume() {
   for (auto ghost : ghosts) {
     // Don't resume move cycle timer unless it is paused.
     if ((ghost->state & ghostState::FRIGHTENED) == 0) {
-      printf("Ghosts::resume_move %s move resume.\n", ghost->name.c_str());
+      //printf("Ghosts::resume_move %s move resume.\n", ghost->name.c_str());
       ghost->move_cycle_timer.start();
     }
   }
@@ -120,7 +120,7 @@ void Ghosts::set_state(uint8_t s) {
 
 void Ghosts::clear_state(uint8_t s) {
   for (auto ghost: ghosts) {
-    printf("Ghosts::clear_state %s state %d\n", ghost->name.c_str(), ghost->state);
+    //printf("Ghosts::clear_state %s state %d\n", ghost->name.c_str(), ghost->state);
     ghost->clear_state(s);
   }
 }
@@ -147,19 +147,6 @@ bool Ghosts::update(uint32_t t) {
 }
 
 void Blinky::init(LevelData ld) {
-  ghostAnims[0] = Rect(0,4,2,2);
-  ghostAnims[1] = Rect(2,4,2,2);
-  ghostAnims[2] = Rect(4,4,2,2);
-  ghostAnims[3] = Rect(6,4,2,2);
-  ghostAnims[4] = Rect(8,4,2,2);
-  ghostAnims[5] = Rect(10,4,2,2);
-  ghostAnims[6] = Rect(12,4,2,2);
-  ghostAnims[7] = Rect(14,4,2,2);
-  ghostAnims[8] = Rect(8,12,2,2);
-  ghostAnims[9] = Rect(8,12,2,2);
-  ghostAnims[11] = Rect(12,12,2,2);
-  ghostAnims[10] = Rect(14,12,2,2);
-
   // When power pill running low sub this in for animation 9.
   //   ghostAnims[9] = Rect(10,12,2,2);
   target_offset = 0;
@@ -176,25 +163,14 @@ void Blinky::init(LevelData ld) {
 
 Blinky::Blinky() {
   name = "Blinky";
+  anim_offset = 0;
+  sprite = anim_offset;
   scatter_target = Point(31 * 8, 1 * 8);
   move_cycle_timer.init(blinky_cycle_timer_callback, 7000, -1);
   init(level_data[0]);
 }
 
 void Pinky::init(LevelData ld) {
-  ghostAnims[0] = Rect(0,8,2,2);
-  ghostAnims[1] = Rect(2,8,2,2);
-  ghostAnims[2] = Rect(4,8,2,2);
-  ghostAnims[3] = Rect(6,8,2,2);
-  ghostAnims[4] = Rect(8,8,2,2);
-  ghostAnims[5] = Rect(10,8,2,2);
-  ghostAnims[6] = Rect(12,8,2,2);
-  ghostAnims[7] = Rect(14,8,2,2);
-  ghostAnims[8] = Rect(8,12,2,2);
-  ghostAnims[9] = Rect(8,12,2,2);
-  ghostAnims[11] = Rect(12,12,2,2);
-  ghostAnims[10] = Rect(14,12,2,2);
-
   target_offset = 4;
   location = Point( (18 * 8), 18 * 8);
   direction = Button::DPAD_LEFT;
@@ -209,25 +185,14 @@ void Pinky::init(LevelData ld) {
 
 Pinky::Pinky() {
   name = "Pinky";
+  anim_offset = 8;
+  sprite = anim_offset;
   scatter_target = Point( (9 * 8), 1 * 8);
   move_cycle_timer.init(pinky_cycle_timer_callback, 7000, -1);
   init(level_data[0]);
 }
 
 void Inky::init(LevelData ld) {
-  ghostAnims[0] = Rect(0,10,2,2);
-  ghostAnims[1] = Rect(2,10,2,2);
-  ghostAnims[2] = Rect(4,10,2,2);
-  ghostAnims[3] = Rect(6,10,2,2);
-  ghostAnims[4] = Rect(8,10,2,2);
-  ghostAnims[5] = Rect(10,10,2,2);
-  ghostAnims[6] = Rect(12,10,2,2);
-  ghostAnims[7] = Rect(14,10,2,2);
-  ghostAnims[8] = Rect(8,12,2,2);
-  ghostAnims[9] = Rect(8,12,2,2);
-  ghostAnims[11] = Rect(12,12,2,2);
-  ghostAnims[10] = Rect(14,12,2,2);
-
   target_offset = 2;
   location = Point( (19 * 8) + 4, 18 * 8);
   direction = Button::DPAD_LEFT;
@@ -242,25 +207,14 @@ void Inky::init(LevelData ld) {
 
 Inky::Inky() {
   name = "Inky";
+  anim_offset = 16;
+  sprite = anim_offset;
   scatter_target = Point( 34 * 8, 37 * 8);
   move_cycle_timer.init(inky_cycle_timer_callback, 7000, -1);
   init(level_data[0]);
 }
 
-void Clyde::init(LevelData ld) {
-  ghostAnims[0] = Rect(0,6,2,2);
-  ghostAnims[1] = Rect(2,6,2,2);
-  ghostAnims[2] = Rect(4,6,2,2);
-  ghostAnims[3] = Rect(6,6,2,2);
-  ghostAnims[4] = Rect(8,6,2,2);
-  ghostAnims[5] = Rect(10,6,2,2);
-  ghostAnims[6] = Rect(12,6,2,2);
-  ghostAnims[7] = Rect(14,6,2,2);
-  ghostAnims[8] = Rect(8,12,2,2);
-  ghostAnims[9] = Rect(8,12,2,2);
-  ghostAnims[11] = Rect(14,12,2,2);
-  ghostAnims[10] = Rect(12,12,2,2);
-  
+void Clyde::init(LevelData ld) {  
   // Wrong but it'll do for now.
   target_offset = -4;
   location = Point( (21 * 8), 18 * 8);
@@ -276,6 +230,8 @@ void Clyde::init(LevelData ld) {
 
 Clyde::Clyde() {
   name = "Clyde";
+  anim_offset = 24;
+  sprite = anim_offset;
   scatter_target = Point( 6 * 8, 37 * 8);
   move_cycle_timer.init(clyde_cycle_timer_callback, 7000, -1);
   init(level_data[0]);