+++ ghosts.hpp
@@ -6,10 +6,10 @@
 #ifndef GHOSTS_H
 #define GHOSTS_H
 
-extern int8_t blinky_cycle_index;
-extern int8_t pinky_cycle_index;
-extern int8_t inky_cycle_index;
-extern int8_t clyde_cycle_index;
+extern uint8_t blinky_cycle_index;
+extern uint8_t pinky_cycle_index;
+extern uint8_t inky_cycle_index;
+extern uint8_t clyde_cycle_index;
 
 extern std::vector<Ghost *> ghosts;